refactor(package): domain package 내 strategy, util, dto 분리
application 내에서 적극적으로 행동하는 주요 domain을 제외, 보조적인 역할이라 판단되는 class들을 용도에 따라 분리
