--- src/main/java/racingcar/domain/MoveStrategy.java
@@ -1,5 +1,0 @@
-package racingcar.domain;
-
-public interface MoveStrategy {
-	boolean isMovable();
-}
