--- src/main/java/racingcar/domain/Random.java
@@ -1,9 +1,0 @@
-package racingcar.domain;
-
-public class Random {
-	private static final java.util.Random RANDOM = new java.util.Random();
-
-	public static int makeRandomValue(int bound) {
-		return RANDOM.nextInt(bound + 1);
-	}
-}
